cloud user storage view expanded to file size modifications

+++ NCloud/NCloud/Services/CloudService.cs
@@ -268,7 +268,7 @@ namespace NCloud.Services
 
                 await Task.Run(() => fi.Delete());
 
-                await UpdateUserStorageUsed(user, (-1)*fileSize);
+                await UpdateUserStorageUsed(user, (-1) * fileSize);
 
                 if (!await SetFileConnectedState(cloudPath, fileName, ChangeOwnerIdentification(ChangeRootName(cloudPath), user.UserName), user, false, false))
                     throw new CloudFunctionStopException("failed to adjust file rights");
@@ -753,17 +753,33 @@ namespace NCloud.Services
             }
         }
 
-        public Task<bool> ModifyFileContent(string file, string content)
+        public async Task<bool> ModifyFileContent(string file, string content, CloudUser user)
         {
+
+            double contentSize = GetStringLengthInBytes(content);
+
+            if (user.UsedSpace + contentSize > user.MaxSpace)
+                throw new CloudFunctionStopException("not enough storage");
+
             try
             {
+                FileInfo fi = new FileInfo(file);
+
+                if (!fi.Exists)
+                    throw new CloudFunctionStopException("File does not exist");
+
+                if (contentSize < fi.Length)
+                    contentSize -= fi.Length;
+
                 File.WriteAllText(ParseRootName(file), content);
 
-                return Task.FromResult<bool>(true);
+                await UpdateUserStorageUsed(user, contentSize);
+
+                return true;
             }
             catch (Exception)
             {
-                return Task.FromResult<bool>(false);
+                return false;
             }
         }
 
@@ -1539,6 +1555,83 @@ namespace NCloud.Services
             return await context.SharedFolders.FirstOrDefaultAsync(x => x.CloudPathFromRoot == cloudPath && x.Name == folderName && x.ConnectedToWeb) != null;
         }
 
+        /// <summary>
+        /// Private method to get a folder size in bytes
+        /// </summary>
+        /// <param name="cloudPath">Path to folder in app</param>
+        /// <returns></returns>
+        /// <exception cref="CloudFunctionStopException">Throws if source folder does not exist</exception>
+        /// <exception cref="Exception">Throws in unexpected error happened</exception>
+        private async Task<double> GetDirectorySize(string cloudPath)
+        {
+            try
+            {
+                DirectoryInfo di = new DirectoryInfo(ParseRootName(cloudPath));
+
+                if (!di.Exists)
+                    throw new CloudFunctionStopException("source directory does not exist");
+
+                Queue<DirectoryInfo> dirData = new Queue<DirectoryInfo>(new DirectoryInfo[] { di });
+
+                double size = 0.0;
+
+                while (dirData.Any())
+                {
+                    DirectoryInfo directory = dirData.Dequeue();
+
+                    foreach (FileInfo fi in directory.GetFiles())
+                    {
+                        if (!fi.Exists)
+                            continue;
+
+                        size += fi.Length;
+                    }
+
+                    foreach (DirectoryInfo dir in directory.GetDirectories())
+                    {
+                        dirData.Enqueue(dir);
+                    }
+                }
+
+                return await Task.FromResult<double>(size);
+            }
+            catch (CloudFunctionStopException ex)
+            {
+                throw new CloudFunctionStopException(ex.Message);
+            }
+            catch (Exception)
+            {
+                throw new Exception("Error while getting directory size");
+            }
+        }
+
+        /// <summary>
+        /// Private method to update storage occupied by user
+        /// </summary>
+        /// <param name="user">Currently logged in user</param>
+        /// <param name="amountToBeAdded">The storage to be added/deducted</param>
+        /// <returns>The modified user</returns>
+        /// <exception cref="CloudFunctionStopException">Throws if user is not found</exception>
+        private async Task<CloudUser> UpdateUserStorageUsed(CloudUser user, double amountToBeAdded)
+        {
+
+            user = await context.Users.FirstOrDefaultAsync(x => x.Id == user.Id) ?? throw new CloudFunctionStopException("user is not found"); //get current user state from database or show error
+
+            user.UsedSpace += amountToBeAdded; //updating user used space
+
+            if (user.UsedSpace < 0)
+                user.UsedSpace = 0;
+
+            if (user.UsedSpace > Constants.UserSpaceSize)
+                user.UsedSpace = Constants.UserSpaceSize;
+
+            context.Users.Update(user);
+
+            await context.SaveChangesAsync();
+
+            return user;
+        }
+
         #endregion
 
         #region Private Static Methods
@@ -1634,80 +1727,13 @@ namespace NCloud.Services
         }
 
         /// <summary>
-        /// Private method to get a folder size in bytes
+        /// Private static method to get a string length in bytes
         /// </summary>
-        /// <param name="cloudPath">Path to folder in app</param>
-        /// <returns></returns>
-        /// <exception cref="CloudFunctionStopException">Throws if source folder does not exist</exception>
-        /// <exception cref="Exception">Throws in unexpected error happened</exception>
-        private async Task<double> GetDirectorySize(string cloudPath)
+        /// <param name="content">The content to measure</param>
+        /// <returns>The size in bytes converted to double</returns>
+        private static double GetStringLengthInBytes(string content)
         {
-            try
-            {
-                DirectoryInfo di = new DirectoryInfo(ParseRootName(cloudPath));
-
-                if (!di.Exists)
-                    throw new CloudFunctionStopException("source directory does not exist");
-
-                Queue<DirectoryInfo> dirData = new Queue<DirectoryInfo>(new DirectoryInfo[] { di });
-
-                double size = 0.0;
-
-                while (dirData.Any())
-                {
-                    DirectoryInfo directory = dirData.Dequeue();
-
-                    foreach (FileInfo fi in directory.GetFiles())
-                    {
-                        if (!fi.Exists)
-                            continue;
-
-                        size += fi.Length;
-                    }
-
-                    foreach (DirectoryInfo dir in directory.GetDirectories())
-                    {
-                        dirData.Enqueue(dir);
-                    }
-                }
-
-                return await Task.FromResult<double>(size);
-            }
-            catch (CloudFunctionStopException ex)
-            {
-                throw new CloudFunctionStopException(ex.Message);
-            }
-            catch (Exception)
-            {
-                throw new Exception("Error while getting directory size");
-            }
-        }
-
-        /// <summary>
-        /// Private method to update storage occupied by user
-        /// </summary>
-        /// <param name="user">Currently logged in user</param>
-        /// <param name="amountToBeAdded">The storage to be added/deducted</param>
-        /// <returns>The modified user</returns>
-        /// <exception cref="CloudFunctionStopException">Throws if user is not found</exception>
-        private async Task<CloudUser> UpdateUserStorageUsed(CloudUser user, double amountToBeAdded)
-        {
-
-            user = await context.Users.FirstOrDefaultAsync(x => x.Id == user.Id) ?? throw new CloudFunctionStopException("user is not found"); //get current user state from database or show error
-
-            user.UsedSpace += amountToBeAdded; //updating user used space
-
-            if (user.UsedSpace < 0)
-                user.UsedSpace = 0;
-
-            if (user.UsedSpace > Constants.UserSpaceSize)
-                user.UsedSpace = Constants.UserSpaceSize;
-
-            context.Users.Update(user);
-
-            await context.SaveChangesAsync();
-
-            return user;
+            return content.Length * sizeof(Char);
         }
 
         #endregion
